requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytz>=2023.3
lightstreamer-client-lib>=1.0.3
